chore(ReportsNew): remove new datahub client service
