+++ src/index.js
@@ -19,8 +19,6 @@ import { HardcodedShopifyClient } from './services/shopify-client.doubles';
 import { getDataHubParams } from './utils';
 import { HardcodedDopplerApiClient } from './services/doppler-api-client.double';
 import { HardcodedIpinfoClient } from './services/ipinfo-client.doubles';
-import { HardcodedDatahubClientNew } from './services/datahub-client-new.doubles';
-
 polyfill();
 
 if (document.querySelector('body').setActive) {
@@ -36,8 +34,6 @@ const forcedServices =
         dopplerLegacyClient: new HardcodedDopplerLegacyClient(),
         dopplerSitesClient: new HardcodedDopplerSitesClient(),
         datahubClient: new HardcodedDatahubClient(),
-        // Temporal until use the new datahub client endpoints
-        datahubClientNew: new HardcodedDatahubClientNew(),
         shopifyClient: new HardcodedShopifyClient(),
         dopplerApiClient: new HardcodedDopplerApiClient(),
         ipinfoClient: new HardcodedIpinfoClient(),
