+++ src/services/pure-di.tsx
@@ -10,7 +10,6 @@ import { DopplerApiClient, HttpDopplerApiClient } from './doppler-api-client';
 import { DopplerSitesClient, HttpDopplerSitesClient } from './doppler-sites-client';
 import { IpinfoClient, HttpIpinfoClient } from './ipinfo-client';
 import { ExperimentalFeatures } from './experimental-features';
-import { DatahubClientNew, HttpDatahubClientNew } from './datahub-client-new';
 
 interface AppConfiguration {
   dopplerLegacyUrl: string;
@@ -34,8 +33,6 @@ export interface AppServices {
   axiosStatic: AxiosStatic;
   appConfiguration: AppConfiguration;
   datahubClient: DatahubClient;
-  // Temporal until use the new datahub client endpoints
-  datahubClientNew: DatahubClientNew;
   dopplerLegacyClient: DopplerLegacyClient;
   sessionManager: SessionManager;
   localStorage: Storage;
@@ -100,19 +97,6 @@ export class AppCompositionRoot implements AppServices {
       'datahubClient',
       () =>
         new HttpDatahubClient({
-          axiosStatic: this.axiosStatic,
-          baseUrl: this.appConfiguration.datahubUrl,
-          connectionDataRef: this.appSessionRef,
-        }),
-    );
-  }
-
-  // Temporal until use the new datahub client endpoints
-  get datahubClientNew() {
-    return this.singleton(
-      'datahubClientNew',
-      () =>
-        new HttpDatahubClientNew({
           axiosStatic: this.axiosStatic,
           baseUrl: this.appConfiguration.datahubUrl,
           connectionDataRef: this.appSessionRef,
